taskmanager-api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pydantic-settings
pydantic[email]
python-multipart

pytest
pytest-asyncio
httpx>=0.24.0
